requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Cryptography for Kalshi API signing
cryptography>=41.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Optional: Interactive CLI menu
inquirer>=3.1.3

# Type hints
typing-extensions>=4.0.0
